+++ docs/integration.rst
@@ -162,7 +162,7 @@ Airflow can be configured to read and write task logs in Azure Blob Storage.
 See :ref:`write-logs-azure`.
 
 Azure CosmosDB 
-'''''''''''''''''' 
+''''''''''''''
  
 AzureCosmosDBHook communicates via the Azure Cosmos library. Make sure that a 
 Airflow connection of type `azure_cosmos` exists. Authorization can be done by supplying a 
@@ -173,20 +173,20 @@ default database and collection to use (see connection `azure_cosmos_default` fo
 - :ref:`AzureCosmosInsertDocumentOperator`: Simple operator to insert document into CosmosDB. 
 - :ref:`AzureCosmosDocumentSensor`: Simple sensor to detect document existence in CosmosDB. 
 
-.. AzureCosmosDBHook: 
- 
+.. _AzureCosmosDBHook:
+
 AzureCosmosDBHook 
-""""""""" 
+"""""""""""""""""
  
 .. autoclass:: airflow.contrib.hooks.azure_cosmos_hook.AzureCosmosDBHook 
  
 AzureCosmosInsertDocumentOperator 
-""""""""" 
+"""""""""""""""""""""""""""""""""
  
-.. autoclass:: airflow.contrib.operators.azure_cosmos_insertdocument_operator.AzureCosmosInsertDocumentOperator 
+.. autoclass:: airflow.contrib.operators.azure_cosmos_operator.AzureCosmosInsertDocumentOperator
  
 AzureCosmosDocumentSensor 
-""""""""" 
+"""""""""""""""""""""""""
  
 .. autoclass:: airflow.contrib.sensors.azure_cosmos_sensor.AzureCosmosDocumentSensor
